// src/app/models/teamMember.model.ts
export class teamMember {
    memberId: number | undefined;
    memberName: string | undefined;
    experience: number | undefined;
    skillset: String[] | undefined;
    currentProfileDescription: string | undefined;
    projectStartDate: string | undefined;
    projectEndDate: string | undefined;
    allocationPercentage: number | undefined;
}